Reorganize files according to the MVC architecture
