+++ schema/mrt.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import List
+
+
+# 捷運站名模型
+class MRTData(BaseModel):
+    data: List[str]
