+++ static/index.html
@@ -14,7 +14,7 @@
       referrerpolicy="no-referrer"
     />
   <!-- hero section的圖片預先加載 -->
-  <link rel="preload" href="/static/image/welcome.png" as="image">
+  <link rel="preload" href="/static/image/banner/welcome.png" as="image">
 </head>
   <title>台北一日遊</title>
 </head>
